move GasOutput to separate file
